Conflicts:
	CMakeLists.txt
	cmake/Modules/FindCMocka.cmake
	src/csync.c
	src/csync_exclude.c
	src/csync_reconcile.c
	src/csync_statedb.c
	src/csync_update.c
	src/csync_util.c
	src/vio/csync_vio.c
	tests/CMakeLists.txt
	tests/csync_tests/check_csync_exclude.c
	tests/csync_tests/check_csync_statedb_load.c
	tests/csync_tests/check_csync_statedb_query.c
	tests/csync_tests/check_csync_update.c
	tests/std_tests/check_std_c_path.c

+++ src/vio/csync_vio_method.h
@@ -30,7 +30,7 @@
 #include "vio/csync_vio_handle.h"
 
 #define VIO_METHOD_HAS_FUNC(method,func) \
-  ((((char *)&((method)->func)) - ((char *)(method)) < (method)->method_table_size) \
+  (((size_t)(((char *)&((method)->func)) - ((char *)(method))) < (method)->method_table_size) \
   && method->func != NULL)
 
 typedef struct csync_vio_method_s csync_vio_method_t;
